+++ days/11/day-11.hs
@@ -177,3 +177,23 @@ main = do
   print "42"
 
   print $ readP_to_S readPMonkeyData f
+
+{-
+  Quick Test:
+
+        ghci> :l day-11.hs 
+        [1 of 1] Compiling Main             ( day-11.hs, interpreted )
+        Ok, one module loaded.
+        ghci> f <- readFile "input-11-test.txt" 
+        ghci> readP_to_S readPMonkeyData f
+        [(ReadPMonkey { id = 0, items = [79,98], rpop = <function> },"  Test: divisible by 23\n    If true: throw to monkey 2\n    If false: throw to monkey 3\n\nMonkey 1:\n  Starting items: 54, 65, 75, 74\n  Operation: new = old + 6\n  Test: divisible by 19\n    If true: throw to monkey 2\n    If false: throw to monkey 0\n\nMonkey 2:\n  Starting items: 79, 60, 97\n  Operation: new = old * old\n  Test: divisible by 13\n    If true: throw to monkey 1\n    If false: throw to monkey 3\n\nMonkey 3:\n  Starting items: 74\n  Operation: new = old + 3\n  Test: divisible by 17\n    If true: throw to monkey 0\n    If false: throw to monkey 1\n")]
+        ghci> :t readP_to_S readPMonkeyData f
+        readP_to_S readPMonkeyData f :: [(ReadPMonkey, String)]
+        ghci> [(rpm, str)] = readP_to_S readPMonkeyData f
+        ghci> rpm
+        ReadPMonkey { id = 0, items = [79,98], rpop = <function> }
+        ghci> rpop rpm 5
+        95
+        ghci> rpop rpm 10
+        190
+  -}